+++ R/utils-httr.R
@@ -1,7 +1,9 @@
 # Adapted from googlesheets package https://github.com/jennybc/googlesheets
 
 # Modifications:
+#  - Updated VERB_n to use the RETRY function from httr
 #  - Added catch_errors() function
+#  - Added build_proxy() function
 
 # Copyright (c) 2017 Jennifer Bryan, Joanna Zhao
 #   
@@ -25,45 +27,33 @@
 
 #' Generic implementation of HTTP methods with retries and authentication
 #' 
-#' @importFrom httr status_code config add_headers
+#' @param verb string; The name of HTTP verb to execute
+#' @importFrom httr RETRY status_code config add_headers
 #' @importFrom stats runif
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-VERB_n <- function(VERB, n = 5) {
-  function(url, headers=character(0), ...) {
-    
+VERB_n <- function(verb) {
+  function(url, headers=character(0), ...) {  
     current_state <- salesforcer_state()
-    for (i in seq_len(n)) {
-      
       if(is.null(current_state$auth_method)){
-        out <- VERB(url = url, add_headers(headers), ...)
+        out <- RETRY(verb = verb, url = url, add_headers(headers), build_proxy(),...)
       } else if(current_state$auth_method == 'OAuth'){
         if(grepl("/services/data/v[0-9]{2}.[0-9]{1}/jobs/ingest", url)){
-          out <- VERB(url = url, add_headers(c(headers, "Authorization"=sprintf("Bearer %s", current_state$token$credentials$access_token))), ...)  
+          out <- RETRY(verb = verb, url = url, add_headers(c(headers, "Authorization"=sprintf("Bearer %s", current_state$token$credentials$access_token))), build_proxy(), ...)  
         } else if(grepl("/services/async", url)){
-          out <- VERB(url = url, add_headers(c(headers, "X-SFDC-Session"=current_state$token$credentials$access_token)), ...)  
+          out <- RETRY(verb = verb, url = url, add_headers(c(headers, "X-SFDC-Session"=current_state$token$credentials$access_token)), build_proxy(), ...)  
         } else {
-          out <- VERB(url = url, config=config(token=current_state$token), add_headers(headers), ...)  
+          out <- RETRY(verb = verb, url = url, config=config(token=current_state$token), add_headers(headers), build_proxy(), ...)  
         }
       } else if (current_state$auth_method == 'Basic') {
         if(grepl("/services/async", url)){
-          out <- VERB(url = url, add_headers(c(headers, "X-SFDC-Session"=current_state$session_id)), ...)
+          out <- RETRY(verb = verb, url = url, add_headers(c(headers, "X-SFDC-Session"=current_state$session_id)), build_proxy(), ...)
         } else {
-          out <- VERB(url = url, add_headers(c(headers, "Authorization"=sprintf("Bearer %s", current_state$session_id))), ...)  
+          out <- RETRY(verb = verb, url = url, add_headers(c(headers, "Authorization"=sprintf("Bearer %s", current_state$session_id))), build_proxy(), ...)  
         }
       }
-      
-      status <- status_code(out)
-      if (status < 500 || i == n) break
-      backoff <- runif(n = 1, min = 0, max = 2 ^ i - 1)
-      ## TO DO: honor a verbose argument or option
-      mess <- paste("HTTP error %s on attempt %d ...\n",
-                    "  backing off %0.2f seconds, retrying")
-      mpf(mess, status, i, backoff)
-      Sys.sleep(backoff)
-    }
-    out
+    return(out)
   }
 }
 
@@ -73,47 +63,35 @@ VERB_n <- function(VERB, n = 5) {
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-rGET <- VERB_n(GET)
+rGET <- VERB_n("GET")
 
 #' POSTs with retries and authentication
 #' 
-#' @importFrom httr POST
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-rPOST <- VERB_n(POST)
+rPOST <- VERB_n("POST")
 
 #' PATCHs with retries and authentication
 #' 
-#' @importFrom httr PATCH
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-rPATCH <- VERB_n(PATCH)
+rPATCH <- VERB_n("PATCH")
 
 #' PUTs with retries and authentication
 #' 
-#' @importFrom httr PUT
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-rPUT <- VERB_n(PUT)
+rPUT <- VERB_n("PUT")
 
 #' DELETEs with retries and authentication
 #' 
-#' @importFrom httr DELETE
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-rDELETE <- VERB_n(DELETE)
-
-## good news: these are handy and call. = FALSE is built-in
-##  bad news: 'fmt' must be exactly 1 string, i.e. you've got to paste, iff
-##             you're counting on sprintf() substitution
-cpf <- function(...) cat(paste0(sprintf(...), "\n"))
-mpf <- function(...) message(sprintf(...))
-wpf <- function(...) warning(sprintf(...), call. = FALSE)
-spf <- function(...) stop(sprintf(...), call. = FALSE)
+rDELETE <- VERB_n("DELETE")
 
 #' Function to catch and print HTTP errors
 #'
@@ -146,4 +124,22 @@ catch_errors <- function(x){
     }
   }
   invisible(FALSE)
+}
+
+#' Function to build a proxy object to pass along with httr requests
+#'
+#' @importFrom httr use_proxy
+#' @note This function is meant to be used internally. Only use when debugging.
+#' @keywords internal
+#' @export
+build_proxy <- function(){
+  proxy <- NULL
+  if (getOption("salesforcer.proxy_url") != ""){
+    proxy <- use_proxy(url = getOption("salesforcer.proxy_url"),
+                       port = as.integer(getOption("salesforcer.proxy_port")),
+                       username = getOption("salesforcer.proxy_username"),
+                       password = getOption("salesforcer.proxy_password"),
+                       auth = getOption("salesforcer.proxy_auth"))
+  }
+  return(proxy)
 }