+++ man-roxygen/batch_size.R
@@ -1,0 +1,2 @@
+#' @param batch_size integer; the number of individual records to be included in a 
+#' single batch uploaded to the Bulk APIs (1.0 or 2.0).